+++ src/CommandLine/Sarif/ArtifactLocation.cs
@@ -1,0 +1,6 @@
+﻿namespace Roslynator.CommandLine.Sarif;
+
+internal class ArtifactLocation
+{
+    public string Uri { get; set; }
+}