+++ src/CommandLine/Sarif/Location.cs
@@ -1,0 +1,6 @@
+﻿namespace Roslynator.CommandLine.Sarif;
+
+internal class Location
+{
+    public PhysicalLocation PhysicalLocation { get; set; }
+}