+++ src/CommandLine/Sarif/Message.cs
@@ -1,0 +1,6 @@
+﻿namespace Roslynator.CommandLine.Sarif;
+
+internal class Message
+{
+    public string Text { get; set; }
+}