+++ src/CommandLine/Sarif/PhysicalLocation.cs
@@ -1,0 +1,7 @@
+﻿namespace Roslynator.CommandLine.Sarif;
+
+internal class PhysicalLocation
+{
+    public ArtifactLocation ArtifactLocation { get; set; }
+    public Region Region { get; set; }
+}